A small bug of initTransforms
Hey guys,
I just found a small bug that the initialization of the transformation from base to camera frame can fail if tf frames are published from a remote computer though the network because of the delay. So I suggest to add a timeout in the lookupTransform().
Later on, there are several lookupTransform() calls in the code, I am not sure whether we should add a duration to all of them.